@@ -9,52 +9,49 @@ use App\Repository\RoleRepository;
 use App\Repository\UserStatusRepository;
 use App\Service\CodeGenerator;
 use App\Service\Mailer;
-use App\Service\Mapper\UserMapper;
 use App\Service\Validator\RegistrationValidator;
+use Doctrine\Common\Annotations\AnnotationReader;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
+use Symfony\Component\Serializer\Mapping\Factory\ClassMetadataFactory;
+use Symfony\Component\Serializer\Mapping\Loader\AnnotationLoader;
+use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+use Symfony\Component\Serializer\Serializer;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class RegistrationService
 {
-    private $passwordHasher;
-    private $validator;
-    private $roleRepository;
-    private $userStatusRepository;
-    private $entityManager;
-    private $mailer;
+    private UserPasswordHasherInterface $passwordHasher;
+    private RoleRepository $roleRepository;
+    private UserStatusRepository $userStatusRepository;
+    private EntityManagerInterface $entityManager;
+    private Mailer $mailer;
 
-    public function __construct(ValidatorInterface  $validator, UserPasswordHasherInterface $passwordHasher,
+    public function __construct(UserPasswordHasherInterface $passwordHasher,
                                 RoleRepository $roleRepository, UserStatusRepository $userStatusRepository,
                                 EntityManagerInterface $entityManager, Mailer $mailer)
     {
         $this->passwordHasher = $passwordHasher;
-        $this->validator = $validator;
         $this->roleRepository = $roleRepository;
         $this->userStatusRepository = $userStatusRepository;
         $this->entityManager = $entityManager;
         $this->mailer = $mailer;
     }
 
-    public function prepare(User $user): array
+    public function prepareUser(array $query): User
     {
-        $response = [];
-        $errors = (new RegistrationValidator())->validateUser($this->validator, $user);
-
-        if(count($errors) > 0) {
-            $response["code"] = 400;
-            $response["messages"] = $errors;
-        }
-
-        return $response;
+        $query = array_intersect_key($query, array_flip(["username", "email", "password"]));
+        return (new Serializer([new ObjectNormalizer()]))
+            ->denormalize($query,"App\Entity\User");
     }
 
     public function register(User $user): array
     {
-        $response = [];
+        $result = [];
 
         $user->setPassword($this->passwordHasher->hashPassword($user, $user->getPassword()));
         $role = $this->roleRepository->find(User::DEFAULT_ROLE_ID);
@@ -70,11 +67,9 @@ class RegistrationService
             $this->mailer->sendConfirmationEmail(
                 $_ENV["FRONTEND_DOMAIN"] . "/confirm?code={$user->getConfirmationCode()}&uid={$user->getId()}",
                 $user);
-            $response["code"] = 200;
         } catch(\Throwable $t) {
-            $response["code"] = 500;
-            $response["message"] = "Failed to send confirmation to email.";
+            $result["message"] = "Failed to send confirmation to email.";
         }
-        return $response;
+        return $result;
     }
 }
